fixed git issues
