--- grp4Calc/BasicMathTest.ctxt
@@ -1,25 +1,0 @@
-#BlueJ class context
-comment0.target=BasicMathTest
-comment0.text=\n\ The\ test\ class\ BasicMathTest.\n\n\ @author\ \ (your\ name)\n\ @version\ (a\ version\ number\ or\ a\ date)\n
-comment1.params=
-comment1.target=BasicMathTest()
-comment1.text=\n\ Default\ constructor\ for\ test\ class\ BasicMathTest\n
-comment10.params=
-comment10.target=void\ testReverse()
-comment2.params=
-comment2.target=void\ testAdd()
-comment3.params=
-comment3.target=void\ testSubtract()
-comment4.params=
-comment4.target=void\ testMultiply()
-comment5.params=
-comment5.target=void\ testDivide()
-comment6.params=
-comment6.target=void\ testSquare()
-comment7.params=
-comment7.target=void\ testSquareRoot()
-comment8.params=
-comment8.target=void\ testExponent()
-comment9.params=
-comment9.target=void\ testInversion()
-numComments=11
