--- grp4Calc/CalcTrig.ctxt
@@ -1,22 +1,0 @@
-#BlueJ class context
-comment0.target=CalcTrig
-comment0.text=\n\ Write\ a\ description\ of\ class\ CalcTrig\ here.\n\n\ @author\ (Shuchi)\n\ @version\ (02/09/2019)\n
-comment1.params=
-comment1.target=CalcTrig()
-comment1.text=\n\ Constructor\ for\ objects\ of\ class\ CalcTrig\n
-comment2.params=num1
-comment2.target=double\ calcSin(double)
-comment2.text=\n\ An\ example\ of\ a\ method\ -\ replace\ this\ comment\ with\ your\ own\n\n\ @param\ \ y\ \ a\ sample\ parameter\ for\ a\ method\n\ @return\ \ \ \ the\ sum\ of\ x\ and\ y\n
-comment3.params=num1
-comment3.target=double\ calcCos(double)
-comment4.params=num1
-comment4.target=double\ calcTan(double)
-comment5.params=num1
-comment5.target=double\ inverseSine(double)
-comment6.params=num1
-comment6.target=double\ inverseCos(double)
-comment7.params=num1
-comment7.target=double\ inverseTan(double)
-comment8.params=
-comment8.target=void\ run()
-numComments=9
