--- grp4Calc/CalcTrigTest.ctxt
@@ -1,25 +1,0 @@
-#BlueJ class context
-comment0.target=CalcTrigTest
-comment0.text=\n\ The\ test\ class\ CalcTrigTest.\n\n\ @author\ \ (Shuchi)\n\ @version\ (02/09/2019)\n
-comment1.params=
-comment1.target=CalcTrigTest()
-comment1.text=\n\ Default\ constructor\ for\ test\ class\ CalcTrigTest\n
-comment2.params=
-comment2.target=void\ CalcSineTest()
-comment3.params=
-comment3.target=void\ CalcCoseTest()
-comment4.params=
-comment4.target=void\ CalcTanTest()
-comment5.params=
-comment5.target=void\ CalcinversesinTest()
-comment6.params=
-comment6.target=void\ CalcinversecosTest()
-comment7.params=
-comment7.target=void\ CalcinversetanTest()
-comment8.params=
-comment8.target=void\ setUp()
-comment8.text=\n\ Sets\ up\ the\ test\ fixture.\n\n\ Called\ before\ every\ test\ case\ method.\n
-comment9.params=
-comment9.target=void\ tearDown()
-comment9.text=\n\ Tears\ down\ the\ test\ fixture.\n\n\ Called\ after\ every\ test\ case\ method.\n
-numComments=10
