--- grp4Calc/Calculator.ctxt
@@ -1,8 +1,0 @@
-#BlueJ class context
-comment0.target=Calculator
-comment1.params=
-comment1.target=Calculator()
-comment1.text=\n\ Constructor\ for\ objects\ of\ class\ Calculator\n
-comment2.params=args
-comment2.target=void\ main(java.lang.String[])
-numComments=3
