--- grp4Calc/Display.ctxt
@@ -1,30 +1,0 @@
-#BlueJ class context
-comment0.target=Display
-comment1.params=
-comment1.target=Display()
-comment10.params=prompt\ args
-comment10.target=java.lang.Long\ getLongInput(java.lang.String,\ java.lang.Object[])
-comment10.text=\n\ @param\ prompt\ \:\ text\ to\ display\ to\ user\n\ @param\ args\ \ \ \:\ optional\ arguments\ to\ send\ for\ string\ formatting\n\ @return\ user's\ input\ as\ long\n
-comment2.params=inputStream\ outputStream
-comment2.target=Display(java.io.InputStream,\ java.io.OutputStream)
-comment3.params=scanner\ printStream
-comment3.target=Display(java.util.Scanner,\ java.io.PrintStream)
-comment4.params=val\ args
-comment4.target=void\ print(java.lang.String,\ java.lang.Object[])
-comment4.text=\n\ @param\ val\ \ \:\ text\ to\ display\ on\ console\n\ @param\ args\ \:\ optional\ arguments\ to\ send\ for\ string\ formatting\n
-comment5.params=val\ args
-comment5.target=void\ println(java.lang.String,\ java.lang.Object[])
-comment5.text=\n\ @param\ val\ \ \:\ text\ to\ display\ on\ console\n\ @param\ args\ \:\ optional\ arguments\ to\ send\ for\ string\ formatting\n
-comment6.params=prompt\ args
-comment6.target=java.lang.String\ getStringInput(java.lang.String,\ java.lang.Object[])
-comment6.text=\n\ @param\ prompt\ \:\ text\ to\ display\ to\ user\n\ @param\ args\ \ \ \:\ optional\ arguments\ to\ send\ for\ string\ formatting\n\ @return\ user's\ input\ as\ String\n
-comment7.params=prompt\ args
-comment7.target=java.lang.Integer\ getIntegerInput(java.lang.String,\ java.lang.Object[])
-comment7.text=\n\ @param\ prompt\ \:\ text\ to\ display\ to\ user\n\ @param\ args\ \ \ \:\ optional\ arguments\ to\ send\ for\ string\ formatting\n\ @return\ user's\ input\ as\ integer\n
-comment8.params=prompt\ args
-comment8.target=java.lang.Double\ getDoubleInput(java.lang.String,\ java.lang.Object[])
-comment8.text=\n\ @param\ prompt\ \:\ text\ to\ display\ to\ user\n\ @param\ args\ \ \ \:\ optional\ arguments\ to\ send\ for\ string\ formatting\n\ @return\ user's\ input\ as\ double\n
-comment9.params=prompt\ args
-comment9.target=java.lang.Float\ getFloatInput(java.lang.String,\ java.lang.Object[])
-comment9.text=\n\ @param\ prompt\ \:\ text\ to\ display\ to\ user\n\ @param\ args\ \ \ \:\ optional\ arguments\ to\ send\ for\ string\ formatting\n\ @return\ user's\ input\ as\ float\n
-numComments=11
