--- grp4Calc/Memory.ctxt
@@ -1,16 +1,0 @@
-#BlueJ class context
-comment0.target=Memory
-comment0.text=\n\ Sets\ memory\ to\ an\ array\ and\ tracks\ and\ controls\ the\ memory\n\n\ @author\ (Davis)\n\ @version\ (Version\ 1.0\ 20190208)\n
-comment1.params=
-comment1.target=Memory()
-comment1.text=\n\ Creates\ an\ array\ with\ size\ 1,\ to\ store\ 1\ memory\ number\n
-comment2.params=input
-comment2.target=void\ getMemory(long)
-comment2.text=\n\ Creats\ a\ method\ that\ sets\ the\ memory\ Array\ to\ what\ is\ currently\ \n\ on\ input\n
-comment3.params=
-comment3.target=void\ getM()
-comment3.text=\n\ Sets\ the\ memory\ back\ to\ 0\n
-comment4.params=
-comment4.target=void\ getMrc()
-comment4.text=\n\ Prints\ what\ is\ in\ memory\ to\ display\n
-numComments=5
