--- grp4Calc/package.bluej
@@ -1,118 +1,0 @@
-#BlueJ package file
-dependency1.from=CalcTrigTest
-dependency1.to=CalcTrig
-dependency1.type=UsesDependency
-dependency2.from=Calculator
-dependency2.to=CalcTrig
-dependency2.type=UsesDependency
-<<<<<<< HEAD
-dependency3.from=Calculator
-dependency3.to=CalcTrig
-dependency3.type=UsesDependency
-dependency4.from=Calculator
-dependency4.to=Display
-dependency4.type=UsesDependency
-dependency5.from=BasicMath
-dependency5.to=Display
-dependency5.type=UsesDependency
-editor.fx.0.height=719
-editor.fx.0.width=1236
-editor.fx.0.x=25
-editor.fx.0.y=30
-=======
-editor.fx.0.height=722
-editor.fx.0.width=800
-editor.fx.0.x=240
-editor.fx.0.y=24
->>>>>>> origin
-objectbench.height=101
-objectbench.width=776
-package.divider.horizontal=0.6
-package.divider.vertical=0.8007380073800738
-package.editor.height=427
-package.editor.width=674
-<<<<<<< HEAD
-package.editor.x=95
-package.editor.y=99
-package.frame.height=600
-package.frame.width=800
-package.numDependencies=5
-package.numTargets=8
-=======
-package.editor.x=68
-package.editor.y=43
-package.frame.height=600
-package.frame.width=800
-package.numDependencies=2
-package.numTargets=3
->>>>>>> origin
-package.showExtends=true
-package.showUses=true
-project.charset=UTF-8
-readme.height=58
-readme.name=@README
-readme.width=47
-readme.x=10
-readme.y=10
-target1.association=CalcTrigTest
-target1.height=50
-target1.name=CalcTrig
-target1.showInterface=false
-target1.type=ClassTarget
-target1.width=80
-target1.x=50
-target1.y=140
-target2.height=50
-target2.name=CalcTrigTest
-target2.showInterface=false
-target2.type=UnitTestTargetJunit4
-target2.width=80
-target2.x=80
-target2.y=110
-target3.height=50
-target3.name=Calculator
-target3.showInterface=false
-target3.type=ClassTarget
-target3.width=90
-<<<<<<< HEAD
-target3.x=180
-target3.y=70
-target4.height=50
-target4.name=CalcTrigTest
-target4.showInterface=false
-target4.type=UnitTestTargetJunit4
-target4.width=100
-target4.x=10
-target4.y=130
-target5.height=50
-target5.name=Memory
-target5.showInterface=false
-target5.type=ClassTarget
-target5.width=80
-target5.x=120
-target5.y=130
-target6.height=50
-target6.name=Calculator
-target6.showInterface=false
-target6.type=ClassTarget
-target6.width=90
-target6.x=373
-target6.y=115
-target7.height=50
-target7.name=BasicMathTest
-target7.showInterface=false
-target7.type=UnitTestTargetJunit4
-target7.width=120
-target7.x=70
-target7.y=10
-target8.height=50
-target8.name=Display
-target8.showInterface=false
-target8.type=ClassTarget
-target8.width=90
-target8.x=120
-target8.y=200
-=======
-target3.x=70
-target3.y=10
->>>>>>> origin
